+++ engine/input/windows/GamepadDeviceXI.cpp
@@ -28,7 +28,7 @@ namespace ouzel::input::windows
             if (result == ERROR_DEVICE_NOT_CONNECTED)
                 throw std::runtime_error{"Gamepad " + std::to_string(playerIndex) + " disconnected"};
             else
-                throw std::system_error{result, std::system_category(), "Failed to get state for gamepad " + std::to_string(playerIndex)};
+                throw std::system_error{static_cast<int>(result), std::system_category(), "Failed to get state for gamepad " + std::to_string(playerIndex)};
         }
 
         if (newState.dwPacketNumber > state.dwPacketNumber)
